Add files via upload
Day by day, energy demand keeps rising so that it is essential to reduce energy consumption for that energy conservation is needed. For conservation of energy the best option is energy audit. Energy audit is a process to determine when, where, why and how energy is used in a plant, industry or a building. Collection of these information helps to identify the situation where there is need to improve energy efficiency and decrease production cost. Normally, an energy audit is carried out by certified energy auditors. By conducting energy audit process in industry, employees begin considering energy as a manageable expense and try to conserve it in day-to-day action. However, Energy auditing can be more efficiently done with the help of an external source or a software with the help of Machine Learning. In this current technology-driven world, Machine Learning is a prominent area which makes our machine or electronic device intelligent. Machine Learning is an efficient way of working with machines which involves training and developing intelligent machines. The main objective of this project is to perform energy auditing through machine learning. In this project we develop and train a system for which, for only some certain defined inputs energy auditing is performed. Finally based upon the user or the consumer requirements the system utilities in a building or in an industry are operated and an efficient way of energy auditing can be observed. The future implementation for this method which involves machine learning is by developing an app or a software which is suitable for installing at system utilities and can be operated efficiently, and this creates an eco-friendly environment which can handle energy conservation and access data about the system utilities efficiently.